Crate unstable_suspension

Computes the suspension map between different unstable Ext groups.

Given an unstable Steenrod module $M$, compute the unstable Ext groups of $\Sigma^k M$ for all $k$ up till the stable range. Each result is printed in the form

n s k: num_gens - matrix

The entries are to be interpreted as follows:

n is the stem, which is defined to be t - s - min_degree

s is the Adams filtration

k is the shift

num_gens is the number of generators in this Ext group

matrix is the matrix representing the suspension map from $\Sigma^k M$. This is omitted if the source or target of the suspension map is trivial, or if they have the same dimension and the matrix is the identity matrix.

The output is best read after sorting with sort -n -k 1 -k 2 -k 3.